+++ app/models/theme.rb
@@ -1,10 +1,12 @@
 class Theme < ApplicationRecord
   before_create :fill_hash_id
+  
+  belongs_to :category
   has_many :essays
 
   has_one_attached :support_material
   
-  validates :description, presence: true
+  validates :description, :category_id, presence: true
   validates :support_material, content_type: 'application/pdf'
 
   private