+++ db/schema.rb
@@ -83,7 +83,7 @@ ActiveRecord::Schema.define(version: 2020_06_25_154609) do
 
   create_table "correction_comments", force: :cascade do |t|
     t.integer "correction_id", null: false
-    t.integer "comment_id", null: false
+    t.integer "comment_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.string "extended_comment"
@@ -179,8 +179,10 @@ ActiveRecord::Schema.define(version: 2020_06_25_154609) do
   create_table "themes", force: :cascade do |t|
     t.string "hash_id", default: "", null: false
     t.string "description"
+    t.integer "category_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.index ["category_id"], name: "index_themes_on_category_id"
     t.index ["hash_id"], name: "index_themes_on_hash_id", unique: true
   end
 
@@ -228,6 +230,7 @@ ActiveRecord::Schema.define(version: 2020_06_25_154609) do
   add_foreign_key "students", "classrooms"
   add_foreign_key "students", "schools"
   add_foreign_key "students", "users"
+  add_foreign_key "themes", "categories"
   add_foreign_key "users", "addresses"
   add_foreign_key "users", "roles"
 end